validation/requirements.txt
opencv-python
torch
clip-by-openai
numpy
mediapipe
Pillow
scipy
ffmpeg-python
librosa
pandas
tqdm